eventapp/backend/requirements.txt
aiocontextvars==0.2.2
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextvars==2.4
cryptography==3.1
Django==3.1
django-cors-headers==3.5.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
dynaconf==3.1.0
hvac==0.10.5
idna==2.10
immutables==0.14
Jinja2==2.11.2
knox==0.1.0
loguru==0.5.1
MarkupSafe==1.1.1
Pillow==7.2.0
pycparser==2.20
pytz==2020.1
requests==2.24.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
win32-setctime==1.0.1
